requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.0.2
flask_cors==3.0.3
pysolr==3.9.0
pytz==2022.7
python-dotenv==0.21.0